Finnally fixed the like and reposts bug.
When a reposts and original posts is at the same page if I click like or repost in one of them. Since there will be 2 like/repost input. My PHP code would Insert the like data 2 times. So there will be 2 like added instead of one.
I fixed this by using a for loop and using the $i to use it as an id for like and repsost button then they won't share a similar class. And I'll use those id to know what post is liked. And because it's an id and unique theres no bug about like 2 likes being inserted instead of one.