Fix lightmode bug
When page is loaded the default mode is gonna be dark. and only changes to light mode after so It's gonna be light then dark. And cause the style to flicker. So I used jQuery's document ready to choose what code I excecute before(set the mode to the localstorage value) and the other stuff to load after